+++ pkg/plugin/datasource.go
@@ -132,10 +132,13 @@ func (d *Datasource) query(ctx context.Context, query backend.DataQuery) backend
 		return newResponseError(err, backend.StatusInternal)
 	}
 
-	frames, err := r.getDataFrames(q)
+	frames, err := r.getDataFrames()
 	if err != nil {
 		err = fmt.Errorf("failed to prepare data from reponse: %w", err)
 		return newResponseError(err, backend.StatusInternal)
+	}
+	for i := range frames {
+		q.addMetadataToMultiFrame(frames[i])
 	}
 
 	return backend.DataResponse{Frames: frames}
