+++ pkg/plugin/response.go
@@ -3,10 +3,7 @@ package plugin
 import (
 	"encoding/json"
 	"fmt"
-	"regexp"
-	"sort"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
@@ -15,10 +12,6 @@ import (
 const (
 	vector, matrix, scalar = "vector", "matrix", "scalar"
 )
-
-const legendFormatAuto = "__auto"
-
-var legendFormatRegexp = regexp.MustCompile(`\{\{\s*(.+?)\s*\}\}`)
 
 // Result represents timeseries from query
 type Result struct {
@@ -47,82 +40,11 @@ type Response struct {
 	Data   Data   `json:"data"`
 }
 
-func addMetadataToMultiFrame(q Query, frame *data.Frame) {
-	if frame.Meta == nil {
-		frame.Meta = &data.FrameMeta{}
-	}
-	if len(frame.Fields) < 2 {
-		return
-	}
-
-	customName := getName(q, frame.Fields[1])
-	if customName != "" {
-		frame.Fields[1].Config = &data.FieldConfig{DisplayNameFromDS: customName}
-	}
-
-	frame.Name = customName
-}
-
-func metricNameFromLabels(f *data.Field) string {
-	labels := f.Labels
-	metricName, hasName := labels["__name__"]
-	numLabels := len(labels) - 1
-	if !hasName {
-		numLabels = len(labels)
-	}
-	labelStrings := make([]string, 0, numLabels)
-	for label, value := range labels {
-		if label != "__name__" {
-			labelStrings = append(labelStrings, fmt.Sprintf("%s=%q", label, value))
-		}
-	}
-
-	switch numLabels {
-	case 0:
-		if hasName {
-			return metricName
-		}
-		return "{}"
-	default:
-		sort.Strings(labelStrings)
-		return fmt.Sprintf("%s{%s}", metricName, strings.Join(labelStrings, ", "))
-	}
-}
-
-func getName(q Query, field *data.Field) string {
-	labels := field.Labels
-	legend := metricNameFromLabels(field)
-
-	if q.LegendFormat == legendFormatAuto {
-		if len(labels) > 0 {
-			legend = ""
-		}
-	} else if q.LegendFormat != "" {
-		result := legendFormatRegexp.ReplaceAllFunc([]byte(q.LegendFormat), func(in []byte) []byte {
-			labelName := strings.Replace(string(in), "{{", "", 1)
-			labelName = strings.Replace(labelName, "}}", "", 1)
-			labelName = strings.TrimSpace(labelName)
-			if val, exists := labels[labelName]; exists {
-				return []byte(val)
-			}
-			return []byte{}
-		})
-		legend = string(result)
-	}
-
-	// If legend is empty brackets, use query expression
-	if legend == "{}" {
-		return q.Expr
-	}
-
-	return legend
-}
-
 type promInstant struct {
 	Result []Result `json:"result"`
 }
 
-func (pi promInstant) dataframes(q Query) (data.Frames, error) {
+func (pi promInstant) dataframes() (data.Frames, error) {
 	frames := make(data.Frames, len(pi.Result))
 	for i, res := range pi.Result {
 		f, err := strconv.ParseFloat(res.Value[1].(string), 64)
@@ -134,7 +56,6 @@ func (pi promInstant) dataframes(q Query) (data.Frames, error) {
 		frames[i] = data.NewFrame("",
 			data.NewField(data.TimeSeriesTimeFieldName, nil, []time.Time{ts}),
 			data.NewField(data.TimeSeriesValueFieldName, data.Labels(res.Labels), []float64{f}))
-		addMetadataToMultiFrame(q, frames[i])
 	}
 
 	return frames, nil
@@ -144,7 +65,7 @@ type promRange struct {
 	Result []Result `json:"result"`
 }
 
-func (pr promRange) dataframes(q Query) (data.Frames, error) {
+func (pr promRange) dataframes() (data.Frames, error) {
 	frames := make(data.Frames, len(pr.Result))
 	for i, res := range pr.Result {
 		timestamps := make([]time.Time, len(res.Values))
@@ -170,7 +91,6 @@ func (pr promRange) dataframes(q Query) (data.Frames, error) {
 		frames[i] = data.NewFrame("",
 			data.NewField(data.TimeSeriesTimeFieldName, nil, timestamps),
 			data.NewField(data.TimeSeriesValueFieldName, data.Labels(res.Labels), values))
-		addMetadataToMultiFrame(q, frames[i])
 	}
 
 	return frames, nil
@@ -194,20 +114,20 @@ func (ps promScalar) dataframes() (data.Frames, error) {
 	return frames, nil
 }
 
-func (r *Response) getDataFrames(q Query) (data.Frames, error) {
+func (r *Response) getDataFrames() (data.Frames, error) {
 	switch r.Data.ResultType {
 	case vector:
 		var pi promInstant
 		if err := json.Unmarshal(r.Data.Result, &pi.Result); err != nil {
 			return nil, fmt.Errorf("umarshal err %s; \n %#v", err, string(r.Data.Result))
 		}
-		return pi.dataframes(q)
+		return pi.dataframes()
 	case matrix:
 		var pr promRange
 		if err := json.Unmarshal(r.Data.Result, &pr.Result); err != nil {
 			return nil, fmt.Errorf("umarshal err %s; \n %#v", err, string(r.Data.Result))
 		}
-		return pr.dataframes(q)
+		return pr.dataframes()
 	case scalar:
 		var ps promScalar
 		if err := json.Unmarshal(r.Data.Result, &ps); err != nil {
